+++ src/features/auth/email-expired/hooks/useEmailVerificationResend.ts
@@ -1,8 +1,9 @@
-import { useResendEmailVerificationMutation } from '@/features/auth/api/authApi'
-import { ApiErrorResponse } from '@/shared/api/api.types'
+import { useResendEmailVerificationMutation } from '@/features/auth'
+import { ApiErrorResponse } from '@/shared/api'
+import { APP_ROUTES } from '@/shared/constant'
+
 import { EmailExpiredInputs } from '../model/schemas/emailExpiredSchema'
 import { useBaseEmailResend } from './useBaseEmailResend'
-import { APP_ROUTES } from '@/shared/constant/app-routes'
 
 export const useEmailVerificationResend = () => {
   const [resendEmailVerification, { isLoading }] = useResendEmailVerificationMutation()
@@ -20,12 +21,14 @@ export const useEmailVerificationResend = () => {
     } catch (error: unknown) {
       if (typeof error === 'object' && error !== null && 'data' in error) {
         const apiError = (error as { data: ApiErrorResponse }).data
+
         if (apiError.messages) {
           apiError.messages.forEach(err => {
             if (err.field === 'email') {
               setError('email', { type: 'custom', message: err.message })
             }
           })
+
           return
         }
       }
